### 2023-04-24 #5404 @editorialbot
[PRE REVIEW]: bayes-toolbox: A Python package for Bayesian statistics
**Submitting author:** @hyosubkim (<a href="http://orcid.org/0000-0003-0109-593X">Hyosub E. Kim</a>)
**Repository:** https://github.com/hyosubkim/bayes-toolbox
**Branch with paper.md** (empty if default branch):
**Version:** v0.1.0
