+++ paper/paper.bib
@@ -1,48 +1,53 @@
 @article{kruschke2018bayesian,
   title={The Bayesian New Statistics: Hypothesis testing, estimation, meta-analysis, and power analysis from a Bayesian perspective},
   author={Kruschke, John K and Liddell, Torrin M},
-  journal={Psychonomic bulletin \& review},
+  journal={Psychonomic Bulletin \& Review},
   volume={25},
   pages={178--206},
   year={2018},
-  publisher={Springer}
+  publisher={Springer},
+  doi = {10.31219/osf.io/ksfyr}
 }
 
 @article{shrout2018psychology,
   title={Psychology, science, and knowledge construction: Broadening perspectives from the replication crisis},
   author={Shrout, Patrick E and Rodgers, Joseph L},
-  journal={Annual review of psychology},
+  journal={Annual Review of Psychology},
   volume={69},
   pages={487--510},
   year={2018},
-  publisher={Annual Reviews}
+  publisher={Annual Reviews},
+  doi = {10.1146/annurev-psych-122216-011845}
 }
 
 @article{patil2010pymc,
   title={PyMC: Bayesian stochastic modelling in Python},
   author={Patil, Anand and Huard, David and Fonnesbeck, Christopher J},
-  journal={Journal of statistical software},
+  journal={Journal of Statistical Software},
   volume={35},
   number={4},
   pages={1},
   year={2010},
-  publisher={Europe PMC Funders}
+  publisher={Europe PMC Funders},
+  doi = {10.18637/jss.v035.i04}
 }
 
 @article{kruschke2014doing,
   title={Doing Bayesian data analysis: A tutorial with R, JAGS, and Stan},
   author={Kruschke, John},
   year={2014},
-  publisher={Academic Press}
+  publisher={Academic Press},
+  doi = {10.1016/c2012-0-00477-2}
 }
 
 @misc{wasserstein2016asa,
-  title={ASA statement on statistical significance and P-values},
-  author={Wasserstein, Ronald L},
-  journal={American Statistician},
+  title={The ASA statement on p-values: Context, Process, and Purpose},
+  author={Wasserstein, Ronald L and Lazar, Nicole A},
+  journal={The American Statistician},
   volume={70},
   number={2},
-  pages={131--133},
+  pages={129--133},
   year={2016},
-  publisher={Amer Statistical Assoc 732 N WASHINGTON ST, ALEXANDRIA, VA 22314-1943 USA}
+  publisher={Taylor \& Francis},
+  doi = {10.1080/00031305.2016.1154108}
 }
