### 2023-04-24 #5404 @hyosubkim
Hi @oliviaguest  - Thanks for your message. I've added the DOIs for all citations now. The [link to the article proof on GitHub](https://github.com/openjournals/joss-papers/blob/joss.05404/joss.05404/10.21105.joss.05404.pdf) reflects the new changes, but when I click to download the article, they're not appearing yet.
