### 2023-04-24 #5404 @hyosubkim
Potential reviewers (not on JOSS list)
---
Alexander Etz (Asst Prof, UT Austin; GitHub handle: EtzAlex, email: alexetz@utexas.edu)
Fernando Irarrázaval (Lead data scientist at Enveritas; GitHub handle: cuchoi)
Taraz Lee (Asst Prof, U of Michigan; email: tarazlee@umich.edu)
Thomas Wiecki (Founder of PyMC Labs; GitHub handle: twiecki)
Ian Greenhouse (Asst Prof, U of Oregon; email: img@uoregon.edu)
Kevin Murphy (Google Research Scientist; GitHub handle: murphyk, email: murphyk@gmail.com)

Potential reviewers from JOSS list [username]
---
alstat
mariobecerra
nhtran93
coatless
ChristopherLucas
szorowi1
