+++ docs/docs/index.md
@@ -8,7 +8,7 @@
 ## Why is this useful?
 
 ### Research
-`bayes-toolbox` provides you with the tools for utilizing and exploring Bayesian statistics in your own research projects right away. By wrapping model definitions of Bayesian generalized linear models into convenient functions, `bayes-toolbox` makes it easy to run Bayesian analyses that are analogous to some of the most commonly used frequentist tests in the behavioral and neural sciences (think t-tests, ANOVAs, regression). All of the models (tests) are executable with one-liners, thereby making them ideal for use in an open, replicable (and Bayesian) workflow (watch this [PyMCon talk](https://www.youtube.com/watch?v=ElfToZ9EBpM) to learn more). **Example use cases and tests for nearly every model are provided in the [`examples`](https://github.com/hyosubkim/bayes-toolbox/tree/main/examples) directory, so you can see what a sensible Bayesian data analysis pipeline looks like.** (Many of the example notebooks are adaptations of [Jordi Warmenhoven's Python/PyMC3 port](https://github.com/JWarmenhoven/DBDA-python) of John Kruschke's excellent textbook ["Doing Bayesian Data Analysis: A Tutorial with R, JAGS, and Stan (DBDA)"](https://sites.google.com/site/doingbayesiandataanalysis/home?authuser=0).) 
+`bayes-toolbox` provides you with the tools for utilizing and exploring Bayesian statistics in your own research projects right away. By wrapping model definitions of Bayesian generalized linear models into convenient functions, `bayes-toolbox` makes it easy to run Bayesian analyses that are analogous to some of the most commonly used frequentist tests in the behavioral and neural sciences (think t-tests, ANOVAs, regression). All of the models (tests) are executable with one-liners, thereby making them ideal for use in an open, replicable (and Bayesian) workflow (watch this [PyMCon talk](https://www.youtube.com/watch?v=ElfToZ9EBpM) to learn more). **Example use cases and tests for nearly every model are provided in the [`examples`](https://github.com/hyosubkim/bayes-toolbox/tree/main/examples) directory, so you can see what a sensible Bayesian data analysis pipeline looks like.** (Many of the example notebooks are adaptations of [Jordi Warmenhoven's Python/PyMC3 port](https://github.com/JWarmenhoven/DBDA-python) of John Kruschke's excellent textbook ["Doing Bayesian Data Analysis: A Tutorial with R, JAGS, and Stan (DBDA)"](https://sites.google.com/site/doingbayesiandataanalysis/home?authuser=0).
 
 The benefits of using `bayes-toolbox` in terms of time and convenience will be most noticeable when utilizing hierarchical (multi-level) models, including the ANOVA-like ones. This is because `bayes-toolbox` takes care of the more finicky steps involved in Bayesian statistical modeling with embedded functions for things like standardizing/unstandardizing variables for more efficient MCMC sampling, parsing categorical variables for easier indexing, and implementing sum-to-zero constraints in ANOVA-like models. These are the sorts of implementational details that can add time (and frustration) when creating an analysis pipeline and discourage otherwise interested scientists from using Bayesian statistics. `bayes-toolbox` now removes those obstacles. 
 
