### 2023-04-26 b2dbaaa
Add random effects model for normal likelihood
