+++ CHANGELOG.MD
@@ -1,0 +1,5 @@
+## Unreleased
+
+Added
+- meta-analysis-binary-outcome-beta-version.py: Random effects meta-analysis model in cases of normal likelihood (4/26/23)
+- meta-analyses.ipynb: Added working example of random effects meta-analysis model (4/26/23)
