### 2023-05-25 #5404 @tbrown122387
> 👋 @alstat, @tbrown122387, @lindemann09 would any of you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html

@samhforbes sorry I don’t have any time at the moment
