### 2023-06-06 #5404 @samhforbes
@ChristopherLucas @BrandonEdwards would any of you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html
