### 2023-06-08 #5526 @ChristopherLucas
@hyosubkim: I've only just started my review, but the Statement of Need, as written, seems like it could do a bit more to explain the package's contribution.

Looking at a few functions, many seem to mostly wrap examples from the PyMC documentation (e.g., the function `BEST()` wraps from [this](https://www.pymc.io/projects/examples/en/latest/case_studies/BEST.html#id49) notebook). The PyMC notebooks are a fantastic pedagogical resources, and the functions defined in `bayes-toolbox` hard-code a lot assumptions that perhaps the user _ought_ to have to specify (can a user define a prior for any of the models implemented in `bayes-toolbox` without editing the source code?).

I also don't agree that being a good Bayesian has anything to do with the replication crisis (which is part of the current justification in the statement of need). In fact, the current justification makes me want to write a comment justifying frequentist statistics, which really doesn't seem like the sort of response a Statement of Need ought to trigger.

The statement of need also mentions the potential pedagogical benefit of the package, but to be honest, at least with the current level of documentation and the degree to which models are hardcoded in `bayes-toolbox`, I'd probably point students to the PyMC docs that you're wrapping, rather than to this package.

I don't mean to sound adversarial - I'm open to being convinced otherwise - but this is my present assessment of the paper and the evaluation of the "substantial scholarly effort" criteria.
