### 2023-06-09 #5526 @hyosubkim
@ChristopherLucas : I appreciate your feedback and hope I can address some of your concerns here. First off, I definitely did NOT interpret your comments as being “adversarial”, but I do hope you’ve had a chance to go beyond the BEST function and notebook, and will examine the many other models provided as well as the documentation hosted [here](https://hyosubkim.github.io/bayes-toolbox/), as I think several of your comments are directly addressed there (e.g., choice of priors, value added by wrapping model functions, source material, etc.). I've also tried to address your initial comments below,  with the hope that I can resolve any misunderstandings and we can collaboratively advance to the next stage of improving the repo and source code.

Getting to one of the major points of your critique, as I understand it, bayes-toolbox was never meant to supplant/replace what the PyMC developers have already provided. Rather, bayes-toolbox is intended to serve as a helpful adjunct to their materials as well as a useful library in its own right for any Python user wanting to execute Bayesian analogues of the most common frequentist tests with one-liners. Wrapping models inside functions, as bayes-toolbox has done, is meant to reduce friction for new users, and lower the bar to entry for those who want to explore Bayesian statistics with Python. Right now, Python users can choose between at least a couple of packages that allow for one-liners to be called in order to run classical/frequentist tests (e.g., Pingouin, SciPy). However, for Bayesian stats, there has only been Bambi, which is excellent, but it does require more advanced knowledge and familiarity with R-brms syntax. Therefore, the goal of bayes-toolbox is to fill an important gap in the Python/Bayesian community, by providing an easy-to-use module for less experienced users that makes it as simple to run Bayesian stats as it is to run frequentist stats. The PyMC developers also recognized this gap and expressed support for bayes-toolbox, which I assume is why I was invited to present on bayes-toolbox at the most [recent PyMC Conference](https://pymcon.com/events/).

As far as wrapping PyMC examples, the main purposes of the BEST notebook provided in the “examples” directory is to illustrate simply the bayes-toolbox syntax (and outputs) and to show that what was previously many lines of code to run the BEST test is now reduced to a one-liner with bayes-toolbox. I explicitly state at the top of the notebook that it was adapted from the PyMC developers’ work. Indeed, much more of the original inspiration and source material for bayes-toolbox is from Jordi Warmenhoven’s Python port of the famous [Kruschke textbook](https://sites.google.com/site/doingbayesiandataanalysis/) written for R users, both of whom I also acknowledge in several places.

Re: priors. Yes, the priors are hard-coded and require (for now) changing the source code if users want something different. This is acknowledged in the “Look before you leap” section of the documentation. The choice of priors is based on what was used in the Kruschke textbook. Thus, for new users or those with “prior paralysis”, bayes-toolbox implements the same diffuse, uninformative priors as Kruschke so as not to strongly influence posterior estimates, and to likely satisfy skeptical reviewers. I figured that users who know enough to change their priors in a principled manner will also find it easy to change the model scaffolding provided by bayes-toolbox or be better served by Bambi.

Re: your comment about whether “being a good Bayesian has anything to do with the replication crisis”, I certainly did not intend to trigger those using frequentist stats (I’m one of them, as I’ve only recently begun migrating towards more Bayesian approaches!). My point was that binary judgments about statistical “significance” or the lack thereof is problematic (as statisticians of all stripes have noted, including the ASA), and that quantifying our uncertainty using Bayesian inference is one very natural way to address this. However, I will gladly edit the language if the other reviewers also had a similar interpretation as you.

In terms of pedagogy, perhaps the question is not whether you would direct your students to the PyMC developers OR bayes-toolbox, but rather if bayes-toolbox, in conjunction with other PyMC-developed materials, would serve them well should they start learning and utilizing Bayesian stats in their own work, especially if they are learning from the Kruschke text and want to use Python, as opposed to R.

Apologies for the long response, but you gave me much to think about and address. I hope I’ve convinced you that bayes-toolbox is not a simple repackaging of what the PyMC developers have already published and that it is a unique scholarly contribution, as it provides a much-needed, streamlined Pythonic implementation of Bayesian stats, not to mention some models that are neither from the PyMC developers' materials or the Kruschke textbook (e.g., meta-analyses).
