### 2023-06-12 #5526 @ChristopherLucas
@hyosubkim That response is incredibly helpful, thanks. I recommend putting a bit of that in the paper and perhaps somewhere in the docs, even, as it helps people know what they're looking at when they find the package.
