### 2023-08-02 #5526 @ChristopherLucas
The examples are generally great, though I noticed that this example references a dataframe that is never loaded: https://hyosubkim.github.io/bayes-toolbox/getting-started/
