### 2023-08-02 #5526 @ChristopherLucas
@hyosubkim What's going on with the tests?

https://github.com/hyosubkim/bayes-toolbox/tree/main/bayes_toolbox/tests
