### 2023-08-02 #5526 @ChristopherLucas
I'm done with my review. I might be missing it, but I think the paper hasn't been revised yet, per @hyosubkim's thorough response to my question above. I'd also add some proper tests before finalizing this package, but other than that, it seems good to go IMO. Great work!
