### 2023-08-02 6880394
Clarify example syntax section
