### 2023-08-02 #5526 @hyosubkim
> The examples are generally great, though I noticed that this example references a dataframe that is never loaded: https://hyosubkim.github.io/bayes-toolbox/getting-started/ (i.e., `df` is referenced but never defined).

@ChristopherLucas - Thanks for pointing this out. I grabbed some lines from the hierarchical regression example and added them to the "Example Syntax" section [here](https://hyosubkim.github.io/bayes-toolbox/getting-started/) to make it clear how to implement bayes-toolbox.
