+++ docs/docs/index.md
@@ -6,20 +6,20 @@
 
 `bayes-toolbox` (aka, Bayesian Statistics Toolbox [BST]) is a Python package for running sophisticated Bayesian analyses in a simple, straight forward manner. 
 
-## Why is this useful?
+## Statement of Need (i.e., Why is this useful?)
 
 ### Research
-`bayes-toolbox` provides you with the tools for utilizing and exploring Bayesian statistics in your own research projects right away. By wrapping model definitions of Bayesian generalized linear models into convenient functions, `bayes-toolbox` makes it easy to run Bayesian analyses that are analogous to some of the most commonly used frequentist tests in the behavioral and neural sciences (think t-tests, ANOVAs, regression). All of the models (tests) are executable with one-liners, thereby making them ideal for use in an open, replicable (and Bayesian) workflow (watch this [PyMCon talk](https://www.youtube.com/watch?v=ElfToZ9EBpM) to learn more). **Example use cases and tests for nearly every model are provided in the [`examples`](https://github.com/hyosubkim/bayes-toolbox/tree/main/examples) directory, so you can see what a sensible Bayesian data analysis pipeline looks like.** (Many of the example notebooks are adaptations of [Jordi Warmenhoven's Python/PyMC3 port](https://github.com/JWarmenhoven/DBDA-python) of John Kruschke's excellent textbook ["Doing Bayesian Data Analysis: A Tutorial with R, JAGS, and Stan (DBDA)"](https://sites.google.com/site/doingbayesiandataanalysis/home?authuser=0).
+`bayes-toolbox` provides you with the tools for utilizing and exploring Bayesian statistics in your own research projects right away. By wrapping model definitions of Bayesian generalized linear models into convenient functions, `bayes-toolbox` makes it easy to run Bayesian analyses that are analogous to some of the most commonly used frequentist tests in the behavioral and neural sciences (think t-tests, ANOVAs, regression). Right now, Python users can choose between several packages that allow for one-liners to be called in order to run classical/frequentist tests (e.g., [Pingouin](https://pingouin-stats.org/build/html/index.html#), [SciPy](https://scipy.org/), [pandas](https://pandas.pydata.org/), [statsmodels](https://www.statsmodels.org/stable/index.html)). In contast, for Bayesian statistics there has been [Bambi](https://bambinos.github.io/bambi/), which is excellent, but it does require more advanced knowledge and familiarity with R-brms syntax. Therefore, the goal of bayes-toolbox is to fill an important gap in the Python/Bayesian community, by providing an easy-to-use module for less experienced users that makes it as simple to run Bayesian stats as it is to run frequentist stats. As all of the models (tests) are executable with one-liners, they are ideal for use in an open, replicable (and Bayesian) workflow (watch this [PyMCon talk](https://www.youtube.com/watch?v=ElfToZ9EBpM) to learn more). **Example use cases and tests for nearly every model are provided in the [`examples`](https://github.com/hyosubkim/bayes-toolbox/tree/main/examples) directory, so you can see what a sensible Bayesian data analysis pipeline looks like.** (Many of the example notebooks are adaptations of [Jordi Warmenhoven's Python/PyMC3 port](https://github.com/JWarmenhoven/DBDA-python) of John Kruschke's excellent textbook ["Doing Bayesian Data Analysis: A Tutorial with R, JAGS, and Stan (DBDA)"](https://sites.google.com/site/doingbayesiandataanalysis/home?authuser=0).
 
[…]
