### 2023-08-02 3b68974
Add more detail to Education
