+++ docs/docs/index.md
@@ -16,7 +16,7 @@ The benefits of using `bayes-toolbox` in terms of time and convenience will be m
 The package is actively being developed in a public [GitHub repository](https://github.com/hyosubkim/bayes-toolbox), and we always welcome new contributors! No contribution is too small. If you have any issues with the code, suggestions on how to improve it, or have requests for additional content, whether new features or tutorials, please [open an issue on Github](https://github.com/hyosubkim/bayes-toolbox/issues). 
 
 ### Education
-`bayes-toolbox` will be very useful if you are going through [Doing Bayesian Data Analysis](https://sites.google.com/site/doingbayesiandataanalysis/home?authuser=0) and want to learn how to implement the models in Python/PyMC. We also highly recommend looking at the [PyMC tutorials](https://www.pymc.io/projects/docs/en/stable/learn/core_notebooks/index.html). 
+`bayes-toolbox` will be very useful if you are going through [Doing Bayesian Data Analysis](https://sites.google.com/site/doingbayesiandataanalysis/home?authuser=0) and want to learn how to implement the models in Python/PyMC. We also highly recommend going through some of the [PyMC tutorials](https://www.pymc.io/projects/docs/en/stable/learn/core_notebooks/index.html) to supplement your understanding. The PyMC developers have also adapted many ideas from the Kruschke text. 
 
 ## Look before you leap
 Please note that the models in `bayes-toolbox` all utilize fairly uninformative, diffuse priors, which are, for the most part, the exact same ones used in the Kruschke text. New users or those with “prior paralysis” will likely be relieved to know that these diffuse, uninformative priors will not exert undue influence over posterior estimates and will likely satisfy skeptical reviewers. However, keep in mind that even though `bayes-toolbox` offers a streamlined interface for performing common statistical tests, the assumptions John Kruschke and we, the developers, make in these models may not be the ones you want to make for your particular question. Therefore, it's a good idea to go through the notebooks in the [`examples`](https://github.com/hyosubkim/bayes-toolbox/tree/main/examples) to make sure the model is appropriate for your applications. Part of the beauty of Bayesian modeling is its flexibility, so if you want to change priors/hyperpriors/etc., feel free to use `bayes-toolbox` as model scaffolding for a new bespoke model fit to your purpose. And consider making it a [contribution](https://hyosubkim.github.io/bayes-toolbox/how-to-contribute/)! You may also want to explore using [Bambi](https://bambinos.github.io/bambi/). 
