### 2023-08-02 #5526 @hyosubkim
> I'm done with my review. I might be missing it, but I think the paper hasn't been revised yet, per @hyosubkim's thorough response to my question above. I'd also add some proper tests before finalizing this package ([this](https://github.com/hyosubkim/bayes-toolbox/tree/main/bayes_toolbox/tests) seems incomplete), but other than that, it seems good to go IMO. Great work!

Thanks, @ChristopherLucas ! Originally, I was going to wait for all three reviews to come in before editing, but I've now gone ahead and added many of your suggested edits into the online [documentation](https://hyosubkim.github.io/bayes-toolbox/). You'll see that the "Statement of Need" and "Education" portions, in particular, have more details and some verbatim quotes from my original response to you. I will now start making similar changes to the paper itself, assuming @samhforbes is in agreement with that plan.
