### 2023-08-02 a8f428a
Edit paper based on JOSS reviews
