+++ paper/paper.bib
@@ -51,3 +51,14 @@
   publisher={Taylor \& Francis},
   doi = {10.1080/00031305.2016.1154108}
 }
+
+@article{wilson2017good,
+  title={Good enough practices in scientific computing},
+  author={Wilson, Greg and Bryan, Jennifer and Cranston, Karen and Kitzes, Justin and Nederbragt, Lex and Teal, Tracy K},
+  journal={PLoS computational biology},
+  volume={13},
+  number={6},
+  pages={e1005510},
+  year={2017},
+  publisher={Public Library of Science}
+}
