### 2023-08-03 b23eadd
Add statement wrt testing and functionality
