### 2023-08-03 #5526 @hyosubkim
@samhforbes - I've attempted to address all of @ChristopherLucas 's comments. In summary, I've fixed the example he pointed out, revised the [paper.md file on GitHub](https://github.com/hyosubkim/bayes-toolbox/blob/main/paper/paper.md) and [online documentation](https://hyosubkim.github.io/bayes-toolbox/) to include the information requested (primarily from this reply https://github.com/openjournals/joss-reviews/issues/5526#issuecomment-1584041847), and clarified my implementation of automatic tests and validating functionality. I'll now wait to hear from you on how to proceed. Thanks!
