### 2023-08-03 #5526 @ChristopherLucas
I appreciate these revisions, @hyosubkim. Even in this revision, I still think the replication crisis discussion is distracting and unnecessary, and that it's difficult to develop your point in a satisfying way given the necessary brevity of the article. That said, I'm happy to sign off on this, especially if @samhforbes and the other reviewers are fine with this.

Great work, congrats on the project.
