### 2023-08-03 #5526 @hyosubkim
Really appreciate your feedback, @ChristopherLucas . I agree with your point about the discussion around the replication crisis. It does seem out of scope for such a short paper and have now removed that entire paragraph (see [here](https://github.com/hyosubkim/bayes-toolbox/blob/main/paper/paper.md). I think it reads better now and focuses more on the positives of the package. Thanks for helping to improve both the software and paper.
