### 2023-08-23 #5526 @hyosubkim
@alstat - Just so you know, I've edited the manuscript based on prior feedback. You can find the latest version [here](https://github.com/hyosubkim/bayes-toolbox/blob/main/paper/paper.md). Thanks.
