### 2023-10-12 #5526 @alstat
Hi @samhforbes, I only have two requests for the package @hyosubkim. Although I am not sure if this is available in the docs, but I can't seem to find it.

1. API documentations - while the functions have doc strings that can be accessed using `help(func)`, I think it is better to have this in the web documentation as well so that users who would want to know the list of APIs can simply go through it in the web, instead of looking into the source and checking which of the functions are exported and can be used.
2. Automated Test - while there is a test folder, I only saw two functions being tested, I'm not sure if this is all of the APIs for the bayes_toolbox. I suggest to use code coverage (see https://about.codecov.io/ or other alternative) if possible to assess how many percentage of the functions have been tested, an 80% codecov is a healthy one.

Let me know if you have questions, @hyosubkim. Thanks for your patience for my sched.
