### 2023-10-14 1d2e4df
Create test.yml, copy from bambi and codecov
