+++ bayes_toolbox/tests/test_glm.py
@@ -1,4 +1,7 @@
+import arviz as az
 import numpy as np
+import pytest
+import pymc as pm
 from bayes_toolbox.glm import standardize, BEST_paired
 
 def test_mus():
@@ -9,12 +12,7 @@ def test_mus():
 def test_sigmas():
     X = np.random.normal(loc=10, scale=20, size=(10,2))
     X_s, _, sigma_X = standardize(X)
-    assert np.allclose(X_s.std(axis=0), np.ones_like(sigma_X), atol=1e-4)
-    
-import pytest
-import numpy as np
-import pymc3 as pm
-from bayes_toolbox.glm import BEST_paired
+    assert np.allclose(X_s.std(axis=0), np.ones_like(sigma_X), atol=1e-4)s
 
 # Define a test for BEST_paired
 def test_BEST_paired():
@@ -28,7 +26,7 @@ def test_BEST_paired():
 
     # Perform assertions to check if the function output is as expected
     assert isinstance(model, pm.Model)
-    assert isinstance(idata, pm.backends.base.InferenceData)
+    assert isinstance(idata, az.data.inference_data.InferenceData)
     assert "mu" in model.named_vars
     assert "sigma" in model.named_vars
     assert "nu_minus_one" in model.named_vars
@@ -36,6 +34,5 @@ def test_BEST_paired():
     assert "likelihood" in model.named_vars
     assert "effect_size" in model.named_vars
 
-# You can add more test cases and assertions as needed.
 
 
