### 2023-10-14 a268976
add tests and update environment.yml
