+++ environment.yml
@@ -29,12 +29,12 @@ dependencies:
   - bzip2=1.0.8=h0d85af4_4
   - c-ares=1.18.1=h0d85af4_0
   - c-compiler=1.5.1=hbf74d83_0
-  - ca-certificates=2022.12.7=h033912b_0
+  - ca-certificates=2023.08.22=hecd8cb5_0
   - cachetools=5.2.0=pyhd8ed1ab_0
   - cairo=1.16.0=h904041c_1014
   - cctools=973.0.1=h76f1dac_10
   - cctools_osx-64=973.0.1=h2b95895_10
-  - certifi=2022.12.7=pyhd8ed1ab_0
+  - certifi=2023.7.22=py311hecd8cb5_0
   - cffi=1.15.1=py311ha86e640_2
   - cftime=1.6.2=py311hd5badaa_1
   - charset-normalizer=2.1.1=pyhd8ed1ab_0
@@ -58,24 +58,25 @@ dependencies:
   - entrypoints=0.4=pyhd8ed1ab_0
   - etuples=0.3.8=pyhd8ed1ab_0
   - executing=1.2.0=pyhd8ed1ab_0
-  - expat=2.5.0=hf0c8a7f_0
+  - expat=2.5.0=hcec6c5f_0
   - fastprogress=1.0.3=pyhd8ed1ab_0
   - filelock=3.8.0=pyhd8ed1ab_0
   - flit-core=3.8.0=pyhd8ed1ab_0
-  - font-ttf-dejavu-sans-mono=2.37=hab24e00_0
-  - font-ttf-inconsolata=3.000=h77eed37_0
-  - font-ttf-source-code-pro=2.038=h77eed37_0
-  - font-ttf-ubuntu=0.83=hab24e00_0
-  - fontconfig=2.14.1=h5bb23bf_0
-  - fonts-conda-ecosystem=1=0
+  - font-ttf-dejavu-sans-mono=2.37=hd3eb1b0_0
+  - font-ttf-inconsolata=2.001=hcb22688_0
+  - font-ttf-source-code-pro=2.030=hd3eb1b0_0
+  - font-ttf-ubuntu=0.83=h8b1ccd4_0
+  - fontconfig=2.14.1=hb0a0c50_2
+  - fonts-anaconda=1=h8fa9717_0
+  - fonts-conda-ecosystem=1=hd3eb1b0_0
   - fonts-conda-forge=1=0
   - fonttools=4.38.0=py311h5547dcb_1
   - freetype=2.12.1=h3f81eb7_0
-  - fribidi=1.0.10=hbcb3906_0
+  - fribidi=1.0.10=haf1e3a3_0
   - gdk-pixbuf=2.42.8=h3648f77_1
   - gettext=0.21.1=h8a4c099_0
   - giflib=5.2.1=hbcb3906_2
-  - graphite2=1.3.13=h2e338ed_1001
+  - graphite2=1.3.14=he9d5cce_1
   - graphviz=6.0.1=ha8464fc_0
   - gtk2=2.24.33=h7c1209e_2
   - gts=0.7.6=hccb3bdf_2
@@ -118,7 +119,7 @@ dependencies:
   - libedit=3.1.20191231=h0678c8f_2
   - libev=4.33=haf1e3a3_1
   - libffi=3.4.2=h0d85af4_5
-  - libgd=2.3.3=h1e214de_3
+  - libgd=2.3.3=he7e2580_1
   - libgfortran=5.0.0=9_5_0_h97931a8_26
   - libgfortran5=11.3.0=h082f757_26
   - libglib=2.74.1=h4c723e1_1
@@ -135,7 +136,7 @@ dependencies:
   - libsqlite=3.39.4=ha978bb4_0
[…]
