+++ bayes_toolbox/tests/test_glm.py
@@ -5,16 +5,19 @@ import pytest
 import pymc as pm
 from bayes_toolbox.glm import *
 
+
 def test_mus():
     X = np.random.normal(loc=10, scale=20, size=(10,2))
     X_s, mu_X, _ = standardize(X)
     assert np.allclose(X_s.mean(axis=0), np.zeros_like(mu_X), atol=1e-4)
     
+
 def test_sigmas():
     X = np.random.normal(loc=10, scale=20, size=(10,2))
     X_s, _, sigma_X = standardize(X)
     assert np.allclose(X_s.std(axis=0), np.ones_like(sigma_X), atol=1e-4)
 
+    
 def test_parse_categorical():
     # Create a sample categorical variable
     x = pd.Series(["A", "B", "A", "C", "B", "C"]).astype("category")
@@ -31,6 +34,33 @@ def test_parse_categorical():
     assert list(levels) == list(expected_levels)
     assert n_levels == expected_n_levels
 
+    
+def test_is_standardized():
+    # Test with a standardized DataFrame
+    X = pd.DataFrame(
+        {'A': ([1.0, 2.0, 3.0]), 
+         'B': ([4.0, 5.0, 6.0])}
+    )
+    X_std, _, _ = standardize(X)
+    assert is_standardized(X_std)
+
+    # Test with a non-standardized DataFrame
+    X_non_std = pd.DataFrame({
+        'A': [1.0, 2.0, 3.0], 
+        'B': [4.0, 5.0, 6.0]}
+    )
+    assert not is_standardized(X_non_std)
+
+    # Test with a standardized single variable
+    single_var = pd.Series([-1.0, -2.0, -3.0])
+    single_var_std, _, _ = standardize(single_var)
+    assert is_standardized(single_var_std)
+
+    # Test with a non-standardized single variable
+    single_var_non_std = pd.Series([1.0, 2.0, 3.0])
+    assert not is_standardized(single_var_non_std)
+
+    
 def test_BEST():
     # Creat test data
     iq_drug = np.array([
@@ -62,7 +92,8 @@ def test_BEST():
     assert "nu" in model.named_vars
     assert "difference of means" in model.named_vars
     assert "effect size" in model.named_vars
-        
+
+    
 def test_BEST_paired():
     # Generate test data
     np.random.seed(0)
@@ -81,6 +112,7 @@ def test_BEST_paired():
     assert "nu" in model.named_vars
     assert "likelihood" in model.named_vars
     assert "effect_size" in model.named_vars
+
     
 def test_robust_linear_regression():
     # Generate test data
@@ -101,6 +133,7 @@ def test_robust_linear_regression():
     assert "nu" in model.named_vars
     assert "nu_log10" in model.named_vars
     assert "likelihood" in model.named_vars
+
     
 def test_hierarchical_regression():
     # Generate test data
@@ -120,6 +153,7 @@ def test_hierarchical_regression():
     assert "zsigma" in model.named_vars
     assert "nu" in model.named_vars
 
+    
 def test_hierarchical_bayesian_anova():
     # Generate test data
     x = pd.Series(['A', 'B', 'A', 'C', 'B', 'C'])
@@ -140,5 +174,86 @@ def test_hierarchical_bayesian_anova():
     assert "b0" in model.named_vars
     assert "b" in model.named_vars
 
+    
+def test_multiple_linear_regression():
+    # Generate test data
+    np.random.seed(0)
+    X = pd.DataFrame({
+        'x1': np.random.normal(0, 1, 100),
+        'x2': np.random.normal(0, 1, 100),
+    })
+    y = 2 * X['x1'] + 3 * X['x2'] + np.random.normal(0, 1, 100)
 
+    # Call the multiple_linear_regression function with test data
[…]
