+++ bayes_toolbox/tests/test_glm.py
@@ -3,6 +3,7 @@ import numpy as np
 import pandas as pd
 import pytest
 import pymc as pm
+import xarray as xr
 from bayes_toolbox.glm import *
 
 
@@ -154,6 +155,56 @@ def test_hierarchical_regression():
     assert "nu" in model.named_vars
 
     
+def test_unstandardize_linreg_parameters():
+    # Test data
+    zbeta0 = 0.5
+    zbeta1 = 0.2
+    sigma = 0.1
+    x = pd.Series([1.0, 2.0, 3.0])
+    y = pd.Series([2.0, 4.0, 6.0])
+
+    # Call the unstandardize_linreg_parameters function with test data
+    beta0, beta1, sigma_unstd = unstandardize_linreg_parameters(zbeta0, zbeta1, sigma, x, y)
+    
+    _, mu_x, sigma_x = standardize(x)
+    _, mu_y, sigma_y = standardize(y)
+    
+    # Calculate expected results
+    expected_beta0 = zbeta0 * sigma_y + mu_y - zbeta1 * mu_x * sigma_y / sigma_x
+    expected_beta1 = zbeta1 * sigma_y / sigma_x
+    expected_sigma = sigma * sigma_y
+
+    # Perform assertions to check if the function output is as expected
+    assert np.isclose(beta0, expected_beta0, atol=1e-5)  # Expected value of beta0
+    assert np.isclose(beta1, expected_beta1, atol=1e-5)  # Expected value of beta1
+    assert np.isclose(sigma_unstd, expected_sigma, atol=1e-5)  # Expected value of sigma
+    
+
+def test_hierarchical_regression():
+    # Generate test data
+    np.random.seed(42)
+    n_subjects = 10
+    n_samples = 100
+    subj = pd.Series(np.repeat(range(n_subjects), n_samples))
+    x = pd.Series(np.random.rand(n_subjects * n_samples))
+    y = pd.Series(2 * x + np.random.randn(n_subjects * n_samples))
+
+    # Run the hierarchical regression
+    model, idata = hierarchical_regression(x, y, subj)
+
+    # Check if the returned InferenceData object contains the necessary variables
+    assert "zbeta0" in idata.posterior
+    assert "zbeta1" in idata.posterior
+    assert "zsigma0" in idata.posterior
+    assert "zsigma1" in idata.posterior
+    assert "zbeta0_s" in idata.posterior
+    assert "zbeta1_s" in idata.posterior
+    assert "zbeta0_s_offset" in idata.posterior
+    assert "zbeta1_s_offset" in idata.posterior
+    assert "nu" in idata.posterior
+    assert "zsigma" in idata.posterior
+
+
 def test_hierarchical_bayesian_anova():
     # Generate test data
     x = pd.Series(['A', 'B', 'A', 'C', 'B', 'C'])
@@ -214,7 +265,54 @@ def test_gamma_shape_rate_from_mode_sd():
     # Perform assertions to check if the function output matches the expected results
     assert rate == pytest.approx(expected_rate, abs=1e-5)
     assert shape == pytest.approx(expected_shape, abs=1e-5)
+    
 
+def test_hierarchical_bayesian_anova():
+    # Generate test data
+    np.random.seed(42)
+    n_samples = 100
+    x = pd.Series(np.random.choice(["A", "B", "C"], n_samples), name="x").astype("category")
+    y = pd.Series(np.random.normal(0, 1, n_samples), name="y")
+
+    # Run hierarchical Bayesian ANOVA
+    model, idata = hierarchical_bayesian_anova(x, y)
+
+    # Check if the returned InferenceData object contains the necessary variables
+    assert "a0" in idata.posterior
+    assert "a" in idata.posterior
+    assert "sigma_a" in idata.posterior
[…]
