+++ bayes_toolbox/tests/test_glm.py
@@ -8,17 +8,17 @@ from bayes_toolbox.glm import *
 
 
 def test_mus():
-    X = np.random.normal(loc=10, scale=20, size=(10,2))
+    X = np.random.normal(loc=10, scale=20, size=(10, 2))
     X_s, mu_X, _ = standardize(X)
     assert np.allclose(X_s.mean(axis=0), np.zeros_like(mu_X), atol=1e-4)
-    
+
 
 def test_sigmas():
-    X = np.random.normal(loc=10, scale=20, size=(10,2))
+    X = np.random.normal(loc=10, scale=20, size=(10, 2))
     X_s, _, sigma_X = standardize(X)
     assert np.allclose(X_s.std(axis=0), np.ones_like(sigma_X), atol=1e-4)
 
-    
+
 def test_parse_categorical():
     # Create a sample categorical variable
     x = pd.Series(["A", "B", "A", "C", "B", "C"]).astype("category")
@@ -35,21 +35,15 @@ def test_parse_categorical():
     assert list(levels) == list(expected_levels)
     assert n_levels == expected_n_levels
 
-    
+
 def test_is_standardized():
     # Test with a standardized DataFrame
-    X = pd.DataFrame(
-        {'A': ([1.0, 2.0, 3.0]), 
-         'B': ([4.0, 5.0, 6.0])}
-    )
+    X = pd.DataFrame({"A": ([1.0, 2.0, 3.0]), "B": ([4.0, 5.0, 6.0])})
     X_std, _, _ = standardize(X)
     assert is_standardized(X_std)
 
     # Test with a non-standardized DataFrame
-    X_non_std = pd.DataFrame({
-        'A': [1.0, 2.0, 3.0], 
-        'B': [4.0, 5.0, 6.0]}
-    )
+    X_non_std = pd.DataFrame({"A": [1.0, 2.0, 3.0], "B": [4.0, 5.0, 6.0]})
     assert not is_standardized(X_non_std)
 
     # Test with a standardized single variable
@@ -61,21 +55,23 @@ def test_is_standardized():
     single_var_non_std = pd.Series([1.0, 2.0, 3.0])
     assert not is_standardized(single_var_non_std)
 
-    
+
 def test_BEST():
-    # Creat test data
+    # Create test data
+    # fmt: off
     iq_drug = np.array([
-    101, 100, 102, 104, 102, 97, 105, 105, 98, 101, 100, 123, 105, 103, 
-    100, 95, 102, 106, 109, 102, 82, 102, 100, 102, 102, 101, 102, 102,
-    103, 103, 97, 97, 103, 101, 97, 104, 96, 103, 124, 101, 101, 100,
-    101, 101, 104, 100, 101
+        101, 100, 102, 104, 102, 97, 105, 105, 98, 101, 100, 123, 105, 103, 
+        100, 95, 102, 106, 109, 102, 82, 102, 100, 102, 102, 101, 102, 102,
+        103, 103, 97, 97, 103, 101, 97, 104, 96, 103, 124, 101, 101, 100,
+        101, 101, 104, 100, 101
     ])
 
     iq_placebo = np.array([
-    99, 101, 100, 101, 102, 100, 97, 101, 104, 101, 102, 102, 100, 105,
-    88, 101, 100, 104, 100, 100, 100, 101, 102, 103, 97, 101, 101, 100,
-    101, 99, 101, 100, 100, 101, 100, 99, 101, 100, 102, 99, 100, 99
+        99, 101, 100, 101, 102, 100, 97, 101, 104, 101, 102, 102, 100, 105,
[…]
