### 2023-10-16 3cb8256
add docstrings to reference.md
