+++ docs/docs/reference.md
@@ -1,0 +1,1 @@
+::: bayes-toolbox.glm
