+++ bayes_toolbox/glm.py
@@ -89,7 +89,7 @@ def is_standardized(X, eps=0.0001):
     """Checks to see if variable is standardized (i.e., N(0, 1)).
 
     Args:
-        x: Random variable.
+        X: Random variable.
         eps: Some small value to represent tolerance level.
 
     Returns:
@@ -116,12 +116,12 @@ def BEST(y, group, n_draws=1000):
         for more details.
 
     Args:
-        y (ndarray/Series): The metric outcome variable.
-        group: The grouping variable providing that indexes into y.
+        y (ndarray/pandas.Series): The metric outcome variable.
+        group (pandas.Series): The grouping variable providing that indexes into y.
         n_draws: Number of random samples to draw from the posterior.
 
     Returns:
-        PyMC Model and InferenceData objects.
+        pymc.Model and arviz.InferenceData objects.
     """
 
     # Convert grouping variable to categorical dtype if it is not already
@@ -192,7 +192,7 @@ def BEST_paired(y1, y2=None, n_draws=1000):
         y2 (ndarray/Series): (Optional) If provided, represents the paired
           sample (i.e., y2 elements are in same order as y1).
     Returns:
-        PyMC Model and InferenceData objects.
+        pymc.Model and arviz.InferenceData objects.
     """
 
     # Check to see if y2 was entered. If so, then this means the
@@ -240,7 +240,7 @@ def robust_linear_regression(x, y, n_draws=1000):
         n_draws: Number of random samples to draw from the posterior.
 
     Returns:
-        PyMC Model and InferenceData objects.
+        pymc.Model and arviz.InferenceData objects.
     """
 
     # Standardize both predictor and outcome variables.
@@ -280,9 +280,9 @@ def unstandardize_linreg_parameters(zbeta0, zbeta1, sigma, x, y):
     Args:
         zbeta0 (): Intercept for standardized data.
         zbeta1 (): Slope for standardized data.
-        mu_y (scalar): Mean of outcome variable.
-        sigma_x (scalar): SD of predictor variable.
-        sigma_y (scalar): SD of outcome variable.
+        sigma: SD of standardized data.
+        x: The predictor (independent) variable.
+        y: The outcome (dependent) variable.
 
     Returns:
         ndarrays
@@ -301,12 +301,12 @@ def hierarchical_regression(x, y, subj, n_draws=1000, acceptance_rate=0.9):
     """A multi-level model for estimating group and individual level parameters.
 
     Args:
-        x: Predictor variable.
-        y: Outcome variable.
-        subj: Subj id variable.
+        x (ndarray/pandas.Series): Predictor variable.
+        y (ndarray/pandas.Series): Outcome variable.
+        subj (pandas.Series): Subj id variable.
 
     Returns:
-        PyMC Model and InferenceData objects.
+        pymc.Model and arviz.InferenceData objects.
     """
     zx, mu_x, sigma_x = standardize(x)
     zy, mu_y, sigma_y = standardize(y)
@@ -355,11 +355,11 @@ def multiple_linear_regression(X, y, n_draws=1000):
     """Perform a Bayesian multiple linear regression.
 
     Args:
-        X (dataframe): Predictor variables are in different columns.
+        X (pandas.DataFrame): Predictor variables are in different columns.
         y (ndarray/Series): The outcome variable.
 
     Returns:
-        PyMC Model and InferenceData objects.
+        pymc.Model and arviz.InferenceData objects.
     """
 
     # Standardize both predictor and outcome variables.
@@ -397,18 +397,17 @@ def multiple_linear_regression(X, y, n_draws=1000):
[…]
