### 2023-10-17 #5526 @hyosubkim
Hi @alstat and @samhforbes . I appreciate your feedback @alstat and have made the changes you requested. Specifically, wrt your two requests:

1. API documentation - I have added documentation for all of the functions on the `bayes_toolbox` webpage (see [here](https://hyosubkim.github.io/bayes-toolbox/reference/)).
2. Automated Testing - I have written full test suites with pytest for the models/functions in `bayes_toolbox` (see [here](https://github.com/hyosubkim/bayes-toolbox/tree/main/bayes_toolbox/tests)). I used [Coverage.py](https://coverage.readthedocs.io/en/7.3.2/) to assess the percentage coverage. Currently, coverage of all models/functions is at 84% (see coverage badges [here](https://hyosubkim.github.io/bayes-toolbox/) and [here](https://github.com/hyosubkim/bayes-toolbox/tree/main)).

@alstat - Thanks for helping to improve `bayes_toolbox`! I hope you're satisfied with the changes I've made. @samhforbes - Please let me know if there's anything else, and I look forward to your editorial assessment!
