### 2023-10-17 #5526 @alstat
@hyosubkim Perfect! The API documentation is well made (you can further improve this in the future by adding brief examples for each API if possible). This makes it easy for new users to understand the extent of the contribution of the package. Thanks! I'm satisfied with the changes and with the coverage of the tests. :)
