+++ docs/docs/reference.md
@@ -1,4 +1,4 @@
-# Reference
+# API Reference
 
 ::: bayes_toolbox.glm
 ::: bayes_toolbox.meta
