+++ pyproject.toml
@@ -25,8 +25,7 @@ dependencies = [
     "numpy~=1.23.4",
     "pandas~=1.5.1",
     "pymc~=4.3.0",
-    "setuptools~=65.5.1",
-    "mkdocstrings-python"
+    "setuptools~=65.5.1"
 ]
 keywords = [
     "probabilistic-programming", 
