### 2023-10-18 c8984f0
Add commands to automate versioning
