+++ bayes_toolbox/__init__.py
@@ -1,0 +1,6 @@
+# file: __init__.py
+
+try:
+    from importlib.metadata import version, PackageNotFoundError
+except ImportError:
+    from importlib_metadata import version, PackageNotFoundError
