+++ pyproject.toml
@@ -1,13 +1,15 @@
 [build-system]
 requires = [
-    "setuptools>=61.0",
-    "wheel >= 0.38",
+    "setuptools>=65",
+    "setuptools_scm[toml]",
+    "wheel",
 ]
 build-backend = "setuptools.build_meta"
 
 [project]
 name = "bayes-toolbox"
-version = "0.1.0"
+# version = "0.1.0"
+dynamic = ["version"]
 authors = [
   { name="Hyosub Kim", email="hyosubkim1@gmail.com" },
 ]
@@ -25,7 +27,8 @@ dependencies = [
     "numpy~=1.23.4",
     "pandas~=1.5.1",
     "pymc~=4.3.0",
-    "setuptools~=65.5.1"
+    "setuptools~=65.5.1",
+    'importlib_metadata; python_version < "3.8"'
 ]
 keywords = [
     "probabilistic-programming", 
@@ -33,6 +36,9 @@ keywords = [
     "bayesian-statistics",
     "statistics"
 ]
+
+[tool.setuptools_scm]
+write_to = "bayes_toolbox/_version.py"
 
 [project.urls]
 "Homepage" = "https://github.com/hyosubkim/bayes-toolbox"
