+++ environment.yml
@@ -192,7 +192,6 @@ dependencies:
   - ptyprocess=0.7.0=pyhd3deb0d_0
   - pure_eval=0.2.2=pyhd8ed1ab_0
   - pycparser=2.21=pyhd8ed1ab_0
-  - pygments=2.13.0=pyhd8ed1ab_0
   - pymc=4.3.0=hd8ed1ab_0
   - pymc-base=4.3.0=pyhd8ed1ab_0
   - pyopenssl=22.1.0=pyhd8ed1ab_0
@@ -251,10 +250,12 @@ dependencies:
     - build==0.10.0
     - causalgraphicalmodels==0.0.4
     - click==8.1.3
+    - colorama==0.4.6
     - coverage==7.3.2
     - docopt==0.6.2
     - docutils==0.19
     - ghp-import==2.1.0
+    - griffe==0.36.7
     - iniconfig==2.0.0
     - jaraco-classes==3.2.3
     - keyring==23.13.1
@@ -262,19 +263,28 @@ dependencies:
     - markdown-it-py==2.2.0
     - mdurl==0.1.2
     - mergedeep==1.3.4
-    - mkdocs==1.4.2
+    - mkdocs==1.5.3
+    - mkdocs-autorefs==0.5.0
+    - mkdocs-material==9.4.6
+    - mkdocs-material-extensions==1.3
+    - mkdocstrings==0.23.0
+    - mkdocstrings-python==1.7.3
     - more-itertools==9.1.0
     - mypy-extensions==1.0.0
     - networkx==3.1
+    - paginate==0.5.6
     - pathspec==0.11.1
     - pipreqs==0.4.13
     - pkginfo==1.9.6
     - pluggy==1.0.0
+    - pygments==2.16.1
+    - pymdown-extensions==10.3
     - pyproject-hooks==1.0.0
     - pytest==7.3.1
     - pyyaml==6.0
     - pyyaml-env-tag==0.1
     - readme-renderer==37.3
+    - regex==2023.10.3
     - requests-toolbelt==0.10.1
     - rfc3986==2.0.0
     - rich==13.3.4
