+++ paper/paper.bib
@@ -61,4 +61,5 @@
   pages={e1005510},
   year={2017},
   publisher={Public Library of Science}
+  doi = {10.1371/journal.pcbi.1005510}
 }
