+++ paper/paper.bib
@@ -60,6 +60,6 @@
   number={6},
   pages={e1005510},
   year={2017},
-  publisher={Public Library of Science}
+  publisher={Public Library of Science},
   doi = {10.1371/journal.pcbi.1005510}
 }
