### 2023-10-20 #5526 @hyosubkim
> Hi @hyosubkim our old friend editorialbot is pulling out 6 references from your .bib file. Do you mean to include these in the paper, or are these left over?

Hi @samhforbes - those were holdovers, which I've now removed. editorialbot still seems happy. :)
