+++ paper/paper.bib
@@ -1,5 +1,5 @@
 @article{patil2010pymc,
-  title={PyMC: Bayesian stochastic modelling in Python},
+  title={PyMC: Bayesian {Stochastic Modelling in Python}},
   author={Patil, Anand and Huard, David and Fonnesbeck, Christopher J},
   journal={Journal of Statistical Software},
   volume={35},
@@ -11,7 +11,7 @@
 }
 
 @article{kruschke2014doing,
-  title={Doing Bayesian data analysis: A tutorial with R, JAGS, and Stan},
+  title={{Doing Bayesian Data Analysis: A Tutorial with R, JAGS, and Stan}},
   author={Kruschke, John},
   year={2014},
   publisher={Academic Press},
