### 2023-10-20 ba7f12a
Change kruschke entry to book type
