+++ paper/paper.bib
@@ -10,8 +10,8 @@
   doi = {10.18637/jss.v035.i04}
 }
 
-@article{kruschke2014doing,
-  title={{Doing Bayesian Data Analysis: A Tutorial with R, JAGS, and Stan}},
+@book{kruschke2014doing,
+  title={Doing {Bayesian Data Analysis: A Tutorial with R, JAGS, and Stan}},
   author={Kruschke, John},
   year={2014},
   publisher={Academic Press},
