### 2023-10-20 #5526 @editorialbot
:wave: @openjournals/sbcs-eics, this paper is ready to be accepted and published.
