+++ README.md
@@ -30,7 +30,7 @@ See the [installation instructions](https://hyosubkim.github.io/bayes-toolbox/ge
 If you use `bayes-toolbox` in your work, please cite using one of the following:
 
 ***APA format:***
-- Kim, H. E. (2023). bayes-toolbox (Version 0.0.1) [Computer software]. https://doi.org/10.5281/zenodo.7339667
+- Kim, H. E. (2023). bayes-toolbox (Version 0.1.1) [Computer software]. https://doi.org/10.5281/zenodo.7849408
 
 ***BibTeX format:***
 ```
