### 2023-10-27 #5526 @editorialbot
:wave: @openjournals/sbcs-eics, this paper is ready to be accepted and published.
