### 2023-10-27 #5526 @oliviaguest
Huge thanks to the reviewers @alstat, @ChristopherLucas and editor @samhforbes! ✨  JOSS appreciates your work and effort. ✨ Also, big congratulations to @hyosubkim! 🥳 🍾
